requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
dacite==1.8.1
discord-webhook==1.3.1
docopt-ng==0.9.0
filelock==3.15.4
greenlet==3.0.3
idna==3.7
mutagen==1.47.0
pathvalidate==3.2.0
pika==1.3.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
scdl==2.11.0
six==1.16.0
soundcloud-v2==1.5.2
SQLAlchemy==1.4.52
termcolor==2.4.0
tqdm==4.66.4
urllib3==2.2.2
